Removed duplicate tests and merged tests
