+++ src/System.Net.Requests/tests/System.Net.Requests.Tests.csproj
@@ -50,7 +50,6 @@
     <Compile Include="FtpWebRequestTest.cs" />
     <Compile Include="LoggingTest.cs" />
     <Compile Include="HttpRequestCachePolicyTest.cs" />
-    <Compile Include="HttpWebRequestHeaderTest.cs" />
     <Compile Include="HttpWebResponseHeaderTest.cs" />
   </ItemGroup>
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
